# docs/site/content/en/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Horreum REST API
  description: "Horreum automated change anomaly detection. For more information,\
    \ please see [https://horreum.hyperfoil.io/](https://horreum.hyperfoil.io/)"
  version: "0.13"
...
paths:
...
  /api/test:
    get:
      tags:
      - Test
      description: Retrieve a paginated list of Tests with available count
      operationId: listTests
      parameters:
      - name: roles
        in: query
        description: "__my, __all or a comma delimited  list of roles"
        schema:
          type: string
        example: __my
      - name: limit
        in: query
        description: limit the number of results
        schema:
          format: int32
          type: integer
        example: 20
      - name: page
        in: query
        description: filter by page number of a paginated list of Tests
        schema:
          format: int32
          type: integer
        example: 2
      - name: sort
        in: query
        description: Field name to sort results
        schema:
          default: name
          type: string
        example: name
      - name: direction
        in: query
        description: Sort direction
        schema:
          $ref: '#/components/schemas/SortDirection'
        example: Ascending
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestQueryResult'
    post:
      tags:
      - Test
      description: Create a new test
      operationId: addTest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Test'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
  /api/test/byName/{name}:
    get:
      tags:
      - Test
      description: Retrieve a test by name
      operationId: getByNameOrId
      parameters:
      - name: name
        in: path
        description: Name of test to retrieve
        required: true
        schema:
          type: string
        example: my-comprehensive-benchmark
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
  /api/test/folders:
    get:
      tags:
      - Test
      description: Retrieve a list of all folders
      operationId: folders
      parameters:
      - name: roles
        in: query
        description: "\"__my\", \"__all\" or a comma delimited  list of roles"
        schema:
          type: string
        example: __my
      responses:
        "200":
          description: List of all folders
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example:
              - quarkus
              - ocp-perf-team
  /api/test/import:
    post:
      tags:
      - Test
      description: Import a previously exported Test either as a new Test or to update
        an existing Test
      operationId: importTest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestExport'
        required: true
      responses:
        "201":
          description: Import a new Test or update an existing Test
  /api/test/summary:
    get:
      tags:
      - Test
      description: Retrieve a summary of Tests in a folder
      operationId: summary
      parameters:
      - name: roles
        in: query
        description: "\"__my\", \"__all\" or a comma delimited  list of roles"
        schema:
          type: string
        example: __my
      - name: folder
        in: query
        description: name of the Folder containing the Tests
        schema:
          type: string
        example: My Team Folder
      - name: limit
        in: query
        description: limit the result count
        schema:
          default: 20
          type: integer
        example: 20
      - name: page
        in: query
        description: 'filter by page number of a paginated list of '
        schema:
          default: 1
          type: integer
        example: 1
      - name: direction
        in: query
        description: Sort direction
        schema:
          allOf:
          - $ref: '#/components/schemas/SortDirection'
          - default: Ascending
        example: Ascending
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestListing'
  /api/test/{id}:
    get:
      tags:
      - Test
      description: Retrieve a test by id
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      - name: token
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
    delete:
      tags:
      - Test
      description: Delete a Test by id
      operationId: deleteTest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "204":
          description: No Content
  /api/test/{id}/addToken:
    post:
      tags:
      - Test
      description: "Add a Test API Token for access to provide access to a test data\
        \ for integrated tooling, e.g. reporting services"
      operationId: addToken
      parameters:
      - name: id
        in: path
        description: ID of test to add token to
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestToken'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                format: int32
                type: integer
  /api/test/{id}/export:
    get:
      tags:
      - Test
      operationId: export
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: A Test definition formatted as json
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestExport'
  /api/test/{id}/fingerprint:
    get:
      tags:
      - Test
      description: List all Fingerprints for a Test
      operationId: listFingerprints
      parameters:
      - name: id
        in: path
        description: Test ID to retrieve Fingerprints for
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fingerprints'
  /api/test/{id}/labelValues:
    get:
      tags:
      - Test
      description: List all Label Values for a Test
      operationId: listLabelValues
      parameters:
      - name: id
        in: path
        description: Test ID to retrieve Label Values for
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      - name: filtering
        in: query
        description: Retrieve values for Filtering Labels
        schema:
          default: true
          type: boolean
        example: true
      - name: metrics
        in: query
        description: Retrieve values for Metric Labels
        schema:
          default: true
          type: boolean
        example: false
      - name: filter
        in: query
        description: either a required json sub-document or path expression
        schema:
          default: "{}"
          type: string
        examples:
          object:
            description: json object that must exist in the values object
            value: "{labelName:necessaryValue,...}"
          string:
            description: valid filtering jsonpath that returns null if not found (not
              predicates)
            value: $.count ? (@ < 20 && @ > 10)
      - name: before
        in: query
        description: ISO-like date time string or epoch millis
        schema:
          default: ""
          type: string
        example: 1970-01-01T00:00:00+00:00 or an integer
      - name: after
        in: query
        description: ISO-like date time string or epoch millis
        schema:
          default: ""
          type: string
        example: 1970-01-01T00:00:00+00:00 or an integer
      - name: sort
        in: query
        description: json path to sortable value or start or stop for sorting by time
        schema:
          default: ""
          type: string
        example: $.label or start or stop
      - name: direction
        in: query
        description: either Ascending or Descending
        schema:
          default: Ascending
          type: string
        example: count
      - name: limit
        in: query
        description: the maximum number of results to include
        schema:
          format: int32
          default: 2147483647
          type: integer
        example: 10
      - name: page
        in: query
        description: which page to skip to when using a limit
        schema:
          format: int32
          default: 0
          type: integer
        example: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExportedLabelValues'
  /api/test/{id}/move:
    post:
      tags:
      - Test
      description: Update the folder for a Test. Tests can be moved to different folders
      operationId: updateFolder
      parameters:
      - name: id
        in: path
        description: Test ID to update
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      - name: folder
        in: query
        description: New folder to store the tests
        schema:
          type: string
        example: My Benchmark Folder
      responses:
        "201":
          description: Created
  /api/test/{id}/notifications:
    post:
      tags:
      - Test
      description: "Update notifications for a Test. It is possible to disable notifications\
        \ for a Test, so that no notifications are sent to subscribers"
      operationId: updateNotifications
      parameters:
      - name: id
        in: path
        description: Test ID to update
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      - name: enabled
        in: query
        description: Whether notifications are enabled
        required: true
        schema:
          type: boolean
        example: false
      responses:
        "201":
          description: Created
  /api/test/{id}/recalculate:
    get:
      tags:
      - Test
      description: Get recalculation status for Test
      operationId: getRecalculationStatus
      parameters:
      - name: id
        in: path
        description: Test ID to retrieve recalculation status for
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecalculationStatus'
    post:
      tags:
      - Test
      description: Recalculate Datasets for Test
      operationId: recalculateDatasetsByTest
      parameters:
      - name: id
        in: path
        description: Test ID to recalculate datasets for
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      responses:
        "201":
          description: Created
  /api/test/{id}/revokeToken/{tokenId}:
    delete:
      tags:
      - Test
      description: Revoke a Token defined for a Test
      operationId: dropTestToken
      parameters:
      - name: id
        in: path
        description: Test ID to revoke token
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      - name: tokenId
        in: path
        description: ID of token to revoke
        required: true
        schema:
          format: int32
          type: integer
        example: 202
      responses:
        "204":
          description: No Content
  /api/test/{id}/tokens:
    get:
      tags:
      - Test
      description: A collection of Test Tokens for a given Test
      operationId: tokens
      parameters:
      - name: id
        in: path
        description: ID of test to retrieve list of tokens
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestToken'
  /api/test/{id}/transformers:
    post:
      tags:
      - Test
      description: Update transformers for Test
      operationId: updateTransformers
      parameters:
      - name: id
        in: path
        description: Test ID to retrieve Label Values for
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                format: int32
                type: integer
        required: true
      responses:
        "201":
          description: Created
  /api/test/{id}/updateAccess:
    post:
      tags:
      - Test
      description: Update the Access configuration for a Test
      operationId: updateTestAccess
      parameters:
      - name: id
        in: path
        description: Test ID to revoke token
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      - name: owner
        in: query
        description: Name of the new owner
        required: true
        schema:
          type: string
        example: perf-team
      - name: access
        in: query
        description: New Access level for the Test
        required: true
        schema:
          $ref: '#/components/schemas/Access'
        example: 0
      responses:
        "201":
          description: Created
components:
  schemas:
...
    Fingerprints:
      description: A list of Fingerprints representing one dataset
      type: object
      properties:
        values:
          type: array
          items:
            $ref: '#/components/schemas/FingerprintValue'
...
    SortDirection:
      enum:
      - Ascending
      - Descending
      type: string
...
    TestExport:
      description: Represents a Test with all associated data used for export/import
        operations.
      type: object
      allOf:
      - $ref: '#/components/schemas/Test'
      properties:
        variables:
          description: Array of Variables associated with test
          type: array
          items:
            $ref: '#/components/schemas/Variable'
        missingDataRules:
          description: Array of MissingDataRules associated with test
          type: array
          items:
            $ref: '#/components/schemas/MissingDataRule'
        experiments:
          description: Array of ExperimentProfiles associated with test
          type: array
          items:
            $ref: '#/components/schemas/ExperimentProfile'
        actions:
          description: Array of Actions associated with test
          type: array
          items:
            $ref: '#/components/schemas/Action'
        subscriptions:
          description: Watcher object associated with test
          type: object
          allOf:
          - $ref: '#/components/schemas/Watch'
        datastore:
          description: Datastore associated with test
          type: object
          allOf:
          - $ref: '#/components/schemas/Datastore'
    TestListing:
      required:
      - tests
      - count
      type: object
      properties:
        tests:
          description: Array of Test Summaries
          type: array
          items:
            $ref: '#/components/schemas/TestSummary'
        count:
          format: int64
          description: Number of tests when pagination is ignored
          type: integer
    TestQueryResult:
      required:
      - tests
      - count
      type: object
      properties:
        tests:
          description: Array of Tests
          type: array
          items:
            $ref: '#/components/schemas/Test'
        count:
          format: int64
          description: Count of available tests. This is a count of tests that the
            current user has access to
          type: integer
          example: 64
...
    TestToken:
      required:
      - id
      - value
      - permissions
      - description
      type: object
      properties:
        id:
          format: int32
          description: Unique Token id
          type: integer
          example: 101
        testId:
          format: int32
          description: Test ID to apply Token
          type: integer
          example: 201
        value:
          description: Test value
          type: string
          example: 094678029a2aaf9a2847502273099bb3a1b2338c2b9c618ed09aef0181666e38
        permissions:
          format: int32
          type: integer
        description:
          description: Token description
          type: string
          example: my reporting service token
...